# Improve sodium implementation performance