For the following test script:

```
<?php

require('vendor/autoload.php');

for ($i = 0; $i < 10000000; $i++) {
        \ParagonIE\ConstantTime\Hex::encode('foo');
}
```

Hyperfine reports:

```
hyperfine 'php -d opcache.enable_cli=1 constant_time_encoding/test.php' 'php -d opcache.enable_cli=1 constant_time_encoding_before/test.php'
Benchmark 1: php -d opcache.enable_cli=1 constant_time_encoding/test.php
  Time (mean ± σ):     423.1 ms ±  19.6 ms    [User: 405.6 ms, System: 17.0 ms]
  Range (min … max):   403.7 ms … 475.7 ms    10 runs
 
Benchmark 2: php -d opcache.enable_cli=1 constant_time_encoding_before/test.php
  Time (mean ± σ):     745.7 ms ±  10.3 ms    [User: 725.9 ms, System: 18.8 ms]
  Range (min … max):   731.3 ms … 759.1 ms    10 runs
 
Summary
  php -d opcache.enable_cli=1 constant_time_encoding/test.php ran
    1.76 ± 0.09 times faster than php -d opcache.enable_cli=1 constant_time_encoding_before/test.php
```